--- game/ts/settings.ts
@@ -1,545 +1,0 @@
-import { escapeHTML, formatDate } from "../../shared/shared";
-import { setAPISaveFile, connectApi, getAPI, recalculateSavefileDropdown } from "./api";
-import { createLoadingUi } from "./loading";
-import { resetAllThemes, getConfigBoolean, setConfigBoolean, getConfigString, setConfigString, processBaseUrl, getActiveSaveFile, createNewSaveFile, renameSaveFile, deleteSaveFile, setActiveSaveFile, getAPISaveFiles, uninstallServer, getConfigNumber, setConfigNumber, getInstalledServers, installServer } from "./savefile";
-import { builtInOfficialServers, setActiveServer } from "./server-manager";
-import { animateDialogClose, animateDialogOpen, ConfirmDialog, CustomDialog, PromptDialog } from "./dialog";
-import { getDisplayStatistics } from "./statistics";
-import { decreaseThemePriority, disableTheme, enableTheme, getEnabledThemeList, getThemeList, increaseThemePriority, ThemeEntry, uninstallTheme, updateMountedCss } from "./theme";
-import { addDLCByUrl } from "./dlc-fetch";
-import fileSize from "filesize";
-import { openDevThemeEditor } from "./theme-editor";
-import escapeMarkdown from 'markdown-escape';
-import { playSound, updateMusicVolume } from "./audio";
-import JSZip from "jszip";
-import browserInfo from "./browser-info";
-
-let themeUpdated = false;
-
-declare const $production: string;
-declare const $build_date: string;
-
-function exportToJson(idbDatabase, cb) {
-  const exportObject = {};
-  const objectStoreNamesSet = new Set(idbDatabase.objectStoreNames);
-  const size = objectStoreNamesSet.size;
-  if (size === 0) {
-    cb(null, JSON.stringify(exportObject));
-  } else {
-    const objectStoreNames = Array.from(objectStoreNamesSet) as string[];
-    const transaction = idbDatabase.transaction(
-        objectStoreNames,
-        'readonly'
-    );
-    transaction.onerror = (event) => cb(event, null);
-
-    objectStoreNames.forEach((storeName) => {
-      const allObjects = {};
-      transaction.objectStore(storeName).openCursor().onsuccess = (event) => {
-        const cursor = event.target.result;
-        if (cursor) {
-          allObjects[cursor.key] = cursor.value;
-          cursor.continue();
-        } else {
-          exportObject[storeName] = allObjects;
-          if (
-            objectStoreNames.length ===
-            Object.keys(exportObject).length
-          ) {
-            cb(null, exportObject);
-          }
-        }
-      };
-    });
-  }
-}
-
-export async function InitSettings() {
-  let settingsOpen = false;
-
-  const elemSettingsDialog = document.querySelector('#dialog-settings') as HTMLElement;
-
-  const alwaysSuggest = getConfigBoolean('always-suggest', false);
-  if (alwaysSuggest) {
-    (document.querySelector('#always-suggest') as HTMLInputElement).checked = true;
-  }
-  const showCategoryNames = getConfigBoolean('show-category-names', false);
-  if (showCategoryNames) {
-    (document.querySelector('#show-category-names') as HTMLInputElement).checked = true;
-    document.body.classList.add('config-show-category-names');
-  }
-  const elementScale = getConfigString('element-scale', 'normal');
-  document.body.setAttribute('data-element-scale', elementScale);
-  (document.querySelector('#element-scale') as HTMLInputElement).value = elementScale;
-
-  const playStartupSound = getConfigBoolean('config-play-startup-sound', true);
-  if (playStartupSound) {
-    (document.querySelector('#play-startup-sound') as HTMLInputElement).checked = true;
-  }
-  const playNewsSound = getConfigBoolean('config-play-news-sound', true);
-  if (playNewsSound) {
-    (document.querySelector('#play-news-sound') as HTMLInputElement).checked = true;
-  }
-  const showInternal = getConfigBoolean('config-show-internal-servers', false);
-  if (showInternal) {
-    (document.querySelector('#show-internal-servers') as HTMLInputElement).checked = true;
-  }
-
-  let lastTab = '';
-  async function focus(tab: string) {
-    if(lastTab === 'theme' && tab !== 'theme' && themeUpdated) {
-      await updateMountedCss();
-      themeUpdated = false;
-    }
-    lastTab = tab;
-
-    elemSettingsDialog.querySelectorAll('[data-settings]').forEach((elem) => {
-      elem.classList[elem.getAttribute('data-settings') === tab ? 'add' : 'remove']('selected');
-    });
-    elemSettingsDialog.querySelectorAll('[data-action]').forEach((elem) => {
-      elem.classList[elem.getAttribute('data-action') === tab ? 'add' : 'remove']('selected');
-    });
-
-    if (tab === 'stats') {
-      const content = document.querySelector('.settings-stats');
-      content.innerHTML = '';
-      getDisplayStatistics().then((stats) => {
-        content.innerHTML = `<h2>Statistics for Current Save File</h2><table><tbody>${stats.map(x => `<tr><td><strong>${x[0]}</strong></td><td>${x[1]}</td></tr>`).join('')}</tbody></table>`;
-      });
-    }
-    if (tab === 'reset') {
-      updateStorageEstimation();
-    }
-  }
-
-  document.querySelectorAll('.settings-category,a[data-action]').forEach(elem => {
-    elem.addEventListener('click', async() => {
-      const action = elem.getAttribute('data-action');
-
-      if (action === 'settings-close') {
-        settingsOpen = false;
-        if(themeUpdated) {
-          themeUpdated = false;
-          await updateMountedCss();
-        }
-        animateDialogClose(elemSettingsDialog);
-      } else {
-        playSound('dialog.button')
-        focus(action);
-      }
-    });
-  });
-  document.querySelectorAll('[data-reset-button]').forEach(elem => {
-    elem.addEventListener('click', async() => {
-      const action = elem.getAttribute('data-reset-button');
-
-      if(action === 'build') {
-        if (await ConfirmDialog({
-          title: 'Re-download game bundle?',
-          text: 'This may help fix updates. It will not modify user data.',
-          trueButton: 'Reset'
-        })) {
-          localStorage.setItem('auto_start', 'true');
-          localStorage.removeItem('cache');
-          if (await caches.has('ELEMENTAL')) {
-            caches.delete('ELEMENTAL');
-          }
-          location.reload();
-        }
-      } else if (action === 'themes') {
-        if ((await PromptDialog({
-          title: 'Clear Downloaded Themes?',
-          text: 'Type "Delete Themes" to confirm.'
-        }))?.toLowerCase() === 'delete themes' ) {
-          await resetAllThemes();
-          localStorage.removeItem('themes-enabled');
-          location.reload();
-        }
-      } else if (action === 'all') {
-        if ((await PromptDialog({
-          title: 'Delete Everything?',
-          text: 'Clear Packs, Themes, and ALL PROGRESS on ALL SERVERS? You will be logged out, unlinking your suggestions from this account. This will not remove your created suggestions from the server. Type "Delete All Data" to confirm.'
-        }))?.toLowerCase() === 'delete all data' ) {
-          localStorage.clear();
-          localStorage.setItem('clear_idb', '1');
-          location.reload();
-        }
-      }
-    });
-  });
-  document.querySelector('#dump-everything').addEventListener('click', async() => {
-    var zip = new JSZip();
-    zip.file("readme.txt", [
-      `This is a data dump for Elemental 4`,
-      ``,
-      `Elemental v${require('../../package.json').version}${$production ? '' : ' (Local Build)'}`,
-      `Built on ${$build_date}`,
-      `${browserInfo.browserName} ${browserInfo.fullVersion} on ${browserInfo.osName}`,
-      `Exported on ${formatDate(new Date())}`
-    ].join('\n'));
-
-    await new Promise(resolve => {
-      const request = indexedDB.open("ELEMENTAL");
-      request.onsuccess = function() {
-        exportToJson(this.result, (_, result) => {
-          Object.keys(result).forEach(name => {
-            if(name !== 'local-forage-detect-blob-support') {
-              zip.file(name + '.json', JSON.stringify(result[name], null, 2))
-            }
-          });
-          resolve()
-        });
-      };
-    });
-
-    const local = {};
-    for (let index = 0; index < localStorage.length; index++) {
-      const key = localStorage.key(index);
-      local[key] = localStorage.getItem(key);
-    }
-    zip.file('local_storage.json', JSON.stringify(local, null, 2))
-
-    zip.generateAsync({ type: "blob" })
-      .then(function(blob) {
-        saveAs(blob, `Elem4-${Date.now()}.zip`);
-      });
-  });
-  document.querySelector("[data-action='open-settings']").addEventListener('click', () => {
-    if (!settingsOpen) {
-      if (document.querySelector('.animate-panel')) {
-        document.querySelector('.suggest-close').dispatchEvent(new MouseEvent('click'));
-      }
-      settingsOpen = true;
-      animateDialogOpen(elemSettingsDialog)
-      focus('general');
-    }
-  });
-  document.querySelector("#open-server-settings").addEventListener('click', () => {
-    if (!settingsOpen) {
-      if (document.querySelector('.animate-panel')) {
-        document.querySelector('.suggest-close').dispatchEvent(new MouseEvent('click'));
-      }
-      settingsOpen = true;
-      animateDialogOpen(elemSettingsDialog)
-      focus('server');
-    }
-  });
-  document.querySelector("#server-connect-main").addEventListener('click', () => {
-    connectApi(builtInOfficialServers[0], null, null);
-  });
-  document.querySelector("#show-category-names").addEventListener('click', (ev) => {
-    const v = (ev.currentTarget as HTMLInputElement).checked;
-    setConfigBoolean('show-category-names', v);
-    document.body.classList[v ? 'add' : 'remove']('config-show-category-names');
-  });
-  document.querySelector("#always-suggest").addEventListener('click', (ev) => {
-    const v = (ev.currentTarget as HTMLInputElement).checked;
-    setConfigBoolean('always-suggest', v);
-  });
-  document.querySelector('#play-startup-sound').addEventListener('click', (ev) => {
-    const v = (ev.currentTarget as HTMLInputElement).checked;
-    setConfigBoolean('config-play-startup-sound', v);
-  })
-  document.querySelector('#play-news-sound').addEventListener('click', (ev) => {
-    const v = (ev.currentTarget as HTMLInputElement).checked;
-    setConfigBoolean('config-play-news-sound', v);
-  })
-  document.querySelector('#show-internal-servers').addEventListener('click', (ev) => {
-    const v = (ev.currentTarget as HTMLInputElement).checked;
-    setConfigBoolean('config-show-internal-servers', v);
-    installServer(null, null)
-  })
-
-  document.querySelector("#element-scale").addEventListener('click', (ev) => {
-    const v = (ev.currentTarget as HTMLInputElement).value;
-    setConfigString('element-scale', v);
-    document.body.setAttribute('data-element-scale', v);
-  });
-
-  const themes = getThemeList();
-  const enabled = getEnabledThemeList().concat('elem4_default');
-
-  const activeColumn = document.querySelector('.active-themes .theme-list')
-  const availableColumn = document.querySelector('.available-themes .theme-list')
-
-  enabled.forEach((id) => {
-    const theme = themes.find(x => x.id === id);
-    if (theme) {
-      activeColumn.appendChild(ThemeDOM(theme));
-    }
-  });
-  themes.forEach((theme) => {
-    if (!enabled.includes(theme.id)) {
-      availableColumn.appendChild(ThemeDOM(theme));
-    }
-  })
-
-  document.querySelector('#theme-apply').addEventListener('click', () => {
-    updateMountedCss();
-    themeUpdated = false;
-  });
-  document.querySelector('#theme-add').addEventListener('click', async() => {
-    const text = await PromptDialog({
-      title: 'Add Theme',
-      text: 'Paste the Theme URL or JSON content here.',
-      confirmButton: 'Add Theme',
-      cancelButton: 'Cancel',
-    });
-
-    if (text) {
-      addDLCByUrl(text, 'theme');
-    }
-  });
-  document.querySelector('#server-add').addEventListener('click', async() => {
-    
-    const text = await PromptDialog({
-      title: 'Add Server',
-      text: 'Paste the Server URL here.',
-      confirmButton: 'Add Server',
-      cancelButton: 'Cancel',
-    });
-
-    if (text) {
-      await addDLCByUrl(text, 'server');
-    }
-  });
-  document.querySelector('#theme-browse').addEventListener('click', () => {
-    window.open('/workshop#themes', '', 'width=800,height=600', true);
-  });
-  document.querySelector('#theme-devmode').addEventListener('click', () => {
-    openDevThemeEditor();
-  });
-  document.querySelector('#server-remove').addEventListener('click', async() => {
-    const conf = getAPI().config;
-    if(await ConfirmDialog({
-      title: 'Remove Server',
-      text: `Remove **${escapeMarkdown(conf.name || `Untitled Server (type=${conf.type})`)}**? This will remove all downloaded server data, and your local save files.`
-    })) {
-      await uninstallServer(getAPI().baseUrl);
-    }
-  });
-
-  const serverSelect = document.querySelector('#change-server') as HTMLSelectElement;
-  serverSelect.value = 'internal:change-btn';
-  serverSelect.addEventListener('change', async() => {
-    const baseUrl = serverSelect.value;
-    if (baseUrl && baseUrl !== 'internal:change-btn') {
-      const ui = createLoadingUi();
-      ui.status('Connecting to ' + processBaseUrl(baseUrl), 0)
-      ui.show();
-      setActiveServer(baseUrl);
-      document.querySelector('.suggest-close').dispatchEvent(new MouseEvent('click'));
-      try {
-        await connectApi(baseUrl, null, ui);
-      } catch (error) {
-        await CustomDialog({
-          title: 'Server Connection Error',
-          parts: [
-            `Could not connect to \`${escapeHTML(baseUrl)}\``,
-            `\`${error.stack}\``
-          ]
-        })
-      }
-      ui.dispose();
-    }
-    serverSelect.value = 'internal:change-btn';
-  })
-
-  const saveSelect = document.querySelector('#change-savefile') as HTMLSelectElement
-  saveSelect.addEventListener('change', async() => {
-    const value = saveSelect.value;
-    if(value.startsWith('save:')) {
-      setAPISaveFile(value.slice(5));
-    }
-  })
-  const saveModifySelect = document.querySelector('#modify-savefile') as HTMLSelectElement
-  saveModifySelect.addEventListener('change', async() => {
-    const value = saveModifySelect.value;
-    const api = getAPI();
-    if(value === 'create') {
-      const name = await PromptDialog({
-        title: 'Create Save File',
-        text: 'Enter the savefile name',
-        defaultInput: `Save #${saveSelect.options.length + 1}`,
-        confirmButton: 'Create'
-      });
-      if(name) {
-        await createNewSaveFile(api, name);
-        recalculateSavefileDropdown();
-      }
-    } else if(value === 'rename') {
-      const name = await PromptDialog({
-        title: 'Rename Save File',
-        text: 'Enter the savefile name',
-        defaultInput:`${saveSelect.options[saveSelect.selectedIndex].text}`,
-        confirmButton: 'Rename'
-      });
-      if(name) {
-        await renameSaveFile(api, await getActiveSaveFile(api), name);
-        recalculateSavefileDropdown();
-      }
-    } else if(value === 'delete') {
-      const confirm = await ConfirmDialog({
-        title: 'Delete Save File',
-        text: 'Are you sure you want to delete this save file?',
-        trueButton: 'Delete',
-        falseButton: 'Keep'
-      });
-      if(confirm) {
-        await deleteSaveFile(api, await getActiveSaveFile(api));
-        setActiveSaveFile(api, (await getAPISaveFiles(api))[0].id);
-        recalculateSavefileDropdown();
-      }
-    }
-    saveModifySelect.value = 'title'
-  });
-
-  await updateStorageEstimation();
-
-  const volumeSound = getConfigNumber('volume-sound', 1);
-  const volumeMusic = getConfigNumber('volume-music', 0.5);
-  (document.querySelector('[data-volume-slider="sound"]') as any).value = volumeSound * 100;
-  document.querySelector('[data-volume-display="sound"]').innerHTML = escapeHTML(`${Math.floor(volumeSound * 100)}%`);
-  document.querySelector('[data-volume-slider="sound"]').addEventListener('input', (x) => {
-    setConfigNumber('volume-sound', (x.currentTarget as any).value/100);
-    document.querySelector('[data-volume-display="sound"]').innerHTML = escapeHTML(`${(x.currentTarget as any).value}%`);
-  });
-  (document.querySelector('[data-volume-slider="music"]') as any).value = volumeMusic * 100;
-  document.querySelector('[data-volume-display="music"]').innerHTML = escapeHTML(`${Math.floor(volumeMusic * 100)}%`);
-  document.querySelector('[data-volume-slider="music"]').addEventListener('input', (x) => {
-    setConfigNumber('volume-music', (x.currentTarget as any).value/100);
-    document.querySelector('[data-volume-display="music"]').innerHTML = escapeHTML(`${(x.currentTarget as any).value}%`);
-    updateMusicVolume();
-  })
-}
-
-interface StorageEstimate {
-  quota: number;
-  usage: number;
-  usageDetails?: {
-    indexedDB: number;
-    caches: number;
-    serviceWorkerRegistrations: number;
-  }
-}
-
-export async function updateStorageEstimation() {
-  try {
-    const { quota, usage, usageDetails } = await navigator.storage.estimate() as StorageEstimate
-    document.getElementById('storage-no-anything').style.display = 'none';
-    document.querySelector('#storage-quota-used').innerHTML = escapeHTML(fileSize(usage));
-    document.querySelector('#storage-quota-total').innerHTML = escapeHTML(fileSize(quota));
-    if(usageDetails) {
-      const caches = usageDetails.caches || 0;
-      const indexedDB = usageDetails.indexedDB || 0;
-      const workers = usageDetails.serviceWorkerRegistrations || 0;
-      const totalBreakdown = caches+indexedDB+workers;
-      document.getElementById('storage-bar').innerHTML = [
-        caches > 0 && `<div style='flex:1 1 ${100*(caches/totalBreakdown)}%' class='storage-caches'></div>`,
-        indexedDB > 0 && `<div style='flex:1 1 ${100*(indexedDB/totalBreakdown)}%' class='storage-indexedDB'></div>`,
-        workers > 0 && `<div style='flex:1 1 ${100*(workers/totalBreakdown)}%' class='storage-workers'></div>`,
-      ].filter(Boolean).join('');
-      document.getElementById('storage-breakdown').innerHTML = [
-        caches > 0 && `<tr><td style='width:max-content'><strong>${fileSize(caches)}</strong></td><td style='display:flex;align-items:center;justify-content:center'><span class='inline-storage-icon storage-caches'></span></td><td>Game and Theme Caches</td></tr>`,
-        indexedDB > 0 && `<tr><td style='width:max-content'><strong>${fileSize(indexedDB)}</strong></td><td style='display:flex;align-items:center;justify-content:center'><span class='inline-storage-icon storage-indexedDB'></span></td><td>Element and Config Databases</td></tr>`,
-        workers > 0 && `<tr><td style='width:max-content'><strong>${fileSize(workers)}</strong></td><td style='display:flex;align-items:center;justify-content:center'><span class='inline-storage-icon storage-workers'></span></td><td>Service Worker</td></tr>`,
-      ].filter(Boolean).join('');
-      document.getElementById('storage-no-breakdown').style.display = 'none';
-    } else {
-      document.getElementById('storage-yes-breakdown').style.display = 'none';
-    }
-  } catch (error) {
-    document.getElementById('storage-yes-breakdown').style.display = 'none';
-    document.getElementById('storage-no-breakdown').style.display = 'none';
-  }
-}
-
-function ThemeDOM(theme: ThemeEntry) {
-  const dom = document.createElement('div');
-  dom.setAttribute('data-theme-item', theme.id)
-  dom.classList.add('theme-item');
-  const moveButtons = document.createElement('div');
-  moveButtons.classList.add('theme-move-buttons');
-  if (theme.icon) {
-    moveButtons.setAttribute('style', `background-image:url(${theme.icon});background-size:cover;`)
-  }
-  dom.appendChild(moveButtons);
-
-  const info = document.createElement('div');
-  info.classList.add('theme-info');
-
-  const themeTitle = document.createElement('div')
-  const themeDescription = document.createElement('div')
-  themeTitle.classList.add('theme-title');
-  themeDescription.classList.add('theme-description');
-  themeTitle.innerHTML = escapeHTML(theme.name);
-  themeDescription.innerHTML = escapeHTML(theme.description);
-  info.appendChild(themeTitle);
-  info.appendChild(themeDescription);
-
-  if(theme.id !== 'elem4_default') {
-    moveButtons.innerHTML = `
-      <svg width="60" height="60" viewBox="0 0 60 60" fill="none" xmlns="http://www.w3.org/2000/svg">
-        <rect width="60" height="60" fill="black" fill-opacity="0.4"/>
-        <path class="svg-btn btn-when-activated btn-down" d="M32 32H55L43.5 49L32 32Z" fill="white"/>
-        <path class="svg-btn btn-when-activated btn-up" d="M55 28L32 28L43.5 11L55 28Z" fill="white"/>
-        <path class="svg-btn btn-when-activated btn-deactivate" d="M28 11V49L5 30L28 11Z" fill="white"/>
-        <path class="svg-btn btn-when-deactivated btn-activate" d="M30 49L30 11L53 30L30 49Z" fill="white"/>
-        ${theme.isBuiltIn ? '' : `<path class="svg-btn btn-when-deactivated btn-trash" d="M22 32H18.5L17.5 31H12.5L11.5 32H8V34H22V32ZM9 47C9 47.5304 9.21071 48.0391 9.58579 48.4142C9.96086 48.7893 10.4696 49 11 49H19C19.5304 49 20.0391 48.7893 20.4142 48.4142C20.7893 48.0391 21 47.5304 21 47V35H9V47Z" fill="white"/>`}
-      </svg>
-    `;
-
-    const btnActivate = moveButtons.querySelector('.btn-activate');
-    const btnDeactivate = moveButtons.querySelector('.btn-deactivate');
-    const btnUp = moveButtons.querySelector('.btn-up');
-    const btnDown = moveButtons.querySelector('.btn-down');
-    const btnTrash = moveButtons.querySelector('.btn-trash');
-
-    btnActivate.addEventListener('click', () => {
-      const activeColumn = document.querySelector('.active-themes .theme-list')
-      activeColumn.prepend(dom);
-      enableTheme(theme.id);
-      themeUpdated = true;
-    });
-    btnDeactivate && btnDeactivate.addEventListener('click', () => {
-      const availableColumn = document.querySelector('.available-themes .theme-list')
-      availableColumn.prepend(dom);
-      disableTheme(theme.id);
-      themeUpdated = true;
-    });
-    btnUp.addEventListener('click', () => {
-      decreaseThemePriority(theme.id);
-      dom.parentElement.insertBefore(dom, dom.previousElementSibling);
-      themeUpdated = true;
-    });
-    btnDown.addEventListener('click', () => {
-      increaseThemePriority(theme.id);
-      dom.parentElement.insertBefore(dom, dom.nextElementSibling.nextElementSibling);
-      themeUpdated = true;
-    });
-    btnTrash && btnTrash.addEventListener('click', () => {
-      dom.remove();
-      disableTheme(theme.id);
-      uninstallTheme(theme.id);
-      themeUpdated = true;
-    });
-  }
-
-  dom.appendChild(info);
-
-  return dom
-}
-
-export function addThemeToUI(theme: any) {
-  if (theme && theme.type === 'elemental4:theme') {
-    const find = document.querySelector('[data-theme-item="' + theme.id + '"]');
-    if(find) {
-      const dom = ThemeDOM(theme);
-      find.parentElement.insertBefore(dom, find);
-      find.remove();
-    } else {
-      const activeColumn = document.querySelector('.active-themes .theme-list')
-      activeColumn.prepend(ThemeDOM(theme));
-    }
-  }
-}
